sync-engine/pyproject.toml
[project]
name = "datavault-sync-engine"
version = "0.1.0"
description = "DataVault Notion → Postgres sync (open-source POC)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
  "httpx>=0.27.0",
  "psycopg[binary]>=3.1.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sync_engine"]
